@@ -1,6 +1,6 @@
 {
   "name": "website",
-  "version": "0.1.0",
+  "version": "0.1.1",
   "private": true,
   "scripts": {
     "start": "webpack-dev-server",
@@ -27,39 +27,39 @@
     }
   },
   "dependencies": {
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-router-dom": "^5.2.0",
-    "react-transition-group": "^4.4.1",
-    "styled-components": "^5.1.1",
-    "typescript": "^3.9.7"
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.13.0",
+    "react-transition-group": "^4.4.5",
+    "styled-components": "^6.0.0-rc.3",
+    "typescript": "^5.1.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.11.0",
-    "@babel/preset-env": "^7.11.0",
-    "@babel/preset-react": "^7.10.4",
-    "@babel/preset-typescript": "^7.10.4",
-    "@hot-loader/react-dom": "^16.13.0",
-    "@testing-library/react": "^10.4.7",
-    "@types/jest": "^26.0.8",
-    "@types/react": "^16.9.44",
-    "@types/react-dom": "^16.9.8",
-    "@types/react-router-dom": "^5.1.5",
-    "@types/react-transition-group": "^4.4.0",
-    "@types/styled-components": "^5.1.2",
-    "@types/webpack": "^4.41.21",
-    "@webpack-cli/init": "^0.3.0",
-    "babel-loader": "^8.1.0",
-    "css-loader": "^4.2.0",
-    "html-webpack-plugin": "^4.3.0",
-    "husky": "^4.3.7",
-    "jest": "^26.2.2",
-    "prettier": "^2.2.1",
-    "react-hot-loader": "^4.12.21",
-    "style-loader": "^1.2.1",
-    "ts-loader": "^8.0.2",
-    "webpack": "^4.44.1",
-    "webpack-cli": "^3.3.12",
-    "webpack-dev-server": "^3.11.0"
+    "@babel/core": "^7.22.5",
+    "@babel/preset-env": "^7.22.5",
+    "@babel/preset-react": "^7.22.5",
+    "@babel/preset-typescript": "^7.22.5",
+    "@hot-loader/react-dom": "^17.0.2",
+    "@testing-library/react": "^14.0.0",
+    "@types/jest": "^29.5.2",
+    "@types/react": "^18.2.12",
+    "@types/react-dom": "^18.2.5",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/react-transition-group": "^4.4.6",
+    "@types/styled-components": "^5.1.26",
+    "@types/webpack": "^5.28.1",
+    "@webpack-cli/init": "^1.1.3",
+    "babel-loader": "^9.1.2",
+    "css-loader": "^6.8.1",
+    "html-webpack-plugin": "^5.5.3",
+    "husky": "^8.0.3",
+    "jest": "^29.5.0",
+    "prettier": "^2.8.8",
+    "react-hot-loader": "^4.13.1",
+    "style-loader": "^3.3.3",
+    "ts-loader": "^9.4.3",
+    "webpack": "^5.87.0",
+    "webpack-cli": "^5.1.4",
+    "webpack-dev-server": "^4.15.1"
   }
 }
